email faculty level

+++ universalbackend/approverdecision.php
@@ -14,56 +14,66 @@ $queid = $parts["queid"];
 
 $decis = $parts["decis"];
 
-if ($decis == "ap") {
+$qcheck = "SELECT * FROM queue_table WHERE que_ID = '$queid' AND queue_status = 1";
+$resqch = $conn->query($qcheck);
+$qcount = mysqli_num_rows($resqch);
 
-    $aplgrsql = "UPDATE ledger_table SET
-        queue_status = 2
-        WHERE que_ID = '$queid'";
-    $resaplgr = $conn->query($aplgrsql);
+if ($qcount > 0) {
+    if ($decis == "ap") {
 
-    if ($resaplgr) {
+        $aplgrsql = "UPDATE ledger_table SET
+                queue_status = 2
+                WHERE que_ID = '$queid'";
+        $resaplgr = $conn->query($aplgrsql);
 
-        $apquesql = "UPDATE queue_table SET
-        queue_status = 2
-        WHERE que_ID = '$queid'";
+        if ($resaplgr) {
 
-        $resapque = $conn->query($apquesql);
+            $apquesql = "UPDATE queue_table SET
+                queue_status = 2
+                WHERE que_ID = '$queid'";
 
-        if ($resapque) {
+            $resapque = $conn->query($apquesql);
 
-            // echo "<script type='text/javascript'> alert('Update Successfully') </script>";
-            echo "<script>window.close();</script>";
+            if ($resapque) {
+
+                // echo "<script type='text/javascript'> alert('Update Successfully') </script>";
+                // echo "<script>window.close();</script>";
+                echo "<script type='text/javascript'> location.href='../user/page.html';</script>";
+            } else {
+
+                echo mysqli_error($conn);
+            }
         } else {
 
             echo mysqli_error($conn);
         }
     } else {
 
-        echo mysqli_error($conn);
+        $cclgrsql = "UPDATE ledger_table SET
+            queue_status = 3
+            WHERE que_ID = '$queid'";
+        $rescclgr = $conn->query($cclgrsql);
+
+        if ($rescclgr) {
+
+            $ccquesql = "DELETE FROM queue_table 
+                WHERE que_ID = '$queid'";
+            $resccque = $conn->query($ccquesql);
+
+            if ($resccque) {
+
+                // echo "<script type='text/javascript'> alert('Update Successfully') </script>";
+                // echo "<script>window.close();</script>";
+                echo "<script type='text/javascript'> location.href='../user/page.html';</script>";
+            } else {
+
+                echo mysqli_error($conn);
+            }
+        } else {
+
+            echo mysqli_error($conn);
+        }
     }
 } else {
-
-    $cclgrsql = "UPDATE ledger_table SET
-        queue_status = 3
-        WHERE que_ID = '$queid'";
-    $rescclgr = $conn->query($cclgrsql);
-
-    if ($rescclgr) {
-
-        $ccquesql = "DELETE FROM queue_table 
-        WHERE que_ID = '$queid'";
-        $resccque = $conn->query($ccquesql);
-
-        if ($resccque) {
-
-            // echo "<script type='text/javascript'> alert('Update Successfully') </script>";
-            echo "<script>window.close();</script>";
-        } else {
-
-            echo mysqli_error($conn);
-        }
-    } else {
-
-        echo mysqli_error($conn);
-    }
+    echo "<script type='text/javascript'> location.href='../user/page.html';</script>";
 }
